+++ src/definition/member/mod.rs
@@ -128,6 +128,7 @@ impl Backend {
             content,
             cursor_offset,
             class_loader: &class_loader,
+            resolved_class_cache: Some(&self.resolved_class_cache),
             function_loader: Some(&function_loader),
         };
         let candidates =